+++ src/extension/background.js
@@ -32,8 +32,7 @@ import {
   setDisplay,
   isValidShareURL,
   storeAuthToken,
-  updateAdminAuthHeaderRules,
-  updateProjectAuthorizationHeaderRules,
+  updateAuthHeaderRules,
 } from './utils.js';
 
 /**
@@ -79,12 +78,6 @@ async function getConfigFromTabUrl(tabUrl) {
       ...share,
       ...getGitHubSettings(share.giturl),
       giturl: undefined,
-    };
-  } else if (tabUrl.startsWith(GH_URL)) {
-    // github url
-    return {
-      ...share,
-      ...getGitHubSettings(tabUrl),
     };
   } else {
     try {
@@ -450,14 +443,26 @@ function checkViewDocSource(id) {
 const externalActions = {
   // updates a project's auth token (admin only)
   updateAuthToken: async ({
-    authToken, exp, owner, repo,
+    authToken,
+    siteToken,
+    siteTokenExpiry,
+    owner,
+    repo,
+    exp,
   }, { url }) => {
     let resp = '';
     try {
       if (!url || new URL(url).origin !== 'https://admin.hlx.page') {
         resp = 'unauthorized sender url';
       } else if (authToken !== undefined && owner && repo) {
-        await storeAuthToken(owner, authToken, exp);
+        await storeAuthToken({
+          owner,
+          repo,
+          authToken,
+          authTokenExpiry: exp,
+          siteToken,
+          siteTokenExpiry,
+        });
         resp = 'close';
       }
     } catch (e) {
@@ -681,8 +686,8 @@ const internalActions = {
   chrome.runtime.onMessage.addListener(async ({ deleteAuthToken }, { tab }) => {
     // check if message contains project config and is sent from tab
     if (tab && tab.id && typeof deleteAuthToken === 'object') {
-      const { owner } = deleteAuthToken;
-      await storeAuthToken(owner, '');
+      const { owner, repo } = deleteAuthToken;
+      await storeAuthToken({ owner, repo, authToken: '' });
     }
   });
 
@@ -694,8 +699,7 @@ const internalActions = {
   // });
 
   await updateHelpContent();
-  await updateAdminAuthHeaderRules();
-  await updateProjectAuthorizationHeaderRules();
+  await updateAuthHeaderRules();
   log.info('sidekick extension initialized');
 })();