+++ src/extension/module.js
@@ -2304,6 +2304,8 @@ import sampleRUM from './rum.js';
             includePaths,
             containerId,
             isContainer,
+            isBadge,
+            badgeVariant,
           } = cfg;
           const condition = (s) => {
             let excluded = false;
@@ -2337,7 +2339,20 @@ import sampleRUM from './rum.js';
             custom: true,
             id: id || `custom-plugin-${i}`,
             condition,
-            button: {
+            container: containerId,
+          };
+
+          if (isBadge) {
+            plugin.feature = true;
+            plugin.elements = [{
+              tag: 'span',
+              text: title,
+              attrs: {
+                class: `hlx-sk-badge hlx-sk-badge-${badgeVariant?.toLowerCase() || 'default'}`,
+              },
+            }];
+          } else {
+            plugin.button = {
               text: (titleI18n && titleI18n[lang]) || title || '',
               action: () => {
                 if (url) {
@@ -2425,9 +2440,9 @@ import sampleRUM from './rum.js';
                 fireEvent(sk, `custom:${eventName || id}`);
               },
               isDropdown: isContainer,
-            },
-            container: containerId,
-          };
+            };
+          }
+
           sk.customPlugins[plugin.id] = plugin;
           // check default plugin
           const defaultPlugin = sk.plugins[plugin.id];
@@ -3108,6 +3123,90 @@ import sampleRUM from './rum.js';
   }
 
   /**
+   * Shows the error view.
+   * @private
+   * @param {Sidekick} sk The sidekick
+   * @param {number} code The error code
+   * @param {string} text The error text
+   * @param {string} actionText The action text
+   * @param {Function} action The action
+   */
+  function showErrorView(sk, code, text, actionText, action) {
+    const url = new URL('view/error/error.html', sk.config.scriptRoot);
+    url.searchParams.set('code', code);
+    url.searchParams.set('text', text);
+    url.searchParams.set('action', actionText);
+
+    const viewOverlay = getViewOverlay(sk, true);
+    viewOverlay.classList.add('hlx-sk-error-view');
+    viewOverlay.querySelector('.title').textContent = i18n(sk, 'error_view_description');
+
+    const container = document.body.appendChild(document.createElement('div'));
+    container.classList.add('container');
+
+    const heading = container.appendChild(document.createElement('h2'));
+    heading.classList.add(`code-${code}`);
+    heading.textContent = code;
+    heading.title = code;
+
+    const description = container.appendChild(document.createElement('p'));
+    description.textContent = text;
+
+    const button = container.appendChild(document.createElement('button'));
+    // button.classList.add('accent');
+    button.textContent = actionText;
+    button.title = actionText;
+    button.addEventListener('click', action);
+    viewOverlay.querySelector('.container').replaceWith(container);
+  }
+
+  /**
+   * Handles site errors from the underlying page.
+   * @private
+   * @param {Sidekick} sk The sidekick
+   */
+  function handleSiteError(sk) {
+    if ((sk.location.host.endsWith('.aem.page')
+      || sk.location.host.endsWith('.aem.live')
+      || sk.location.hostname === 'localhost')
+      && !document.querySelector('body > main > div')
+      && document.querySelector('body > pre') === document.body.children[1]) {
+      // 401
+      if (document.querySelector('body > pre').textContent.trim() === '401 Unauthorized') {
+        const loggedIn = sk.isAuthenticated();
+        showErrorView(
+          sk,
+          401,
+          i18n(sk, !loggedIn ? 'error_site_401' : 'error_site_401_relogin'),
+          i18n(sk, !loggedIn ? 'user_login' : 'user_relogin'),
+          () => {
+            login(sk);
+            sk.addEventListener('loggedin', () => {
+              window.location.reload();
+            }, { once: true });
+          },
+        );
+        return;
+      }
+      // 403
+      if (document.querySelector('body > pre').textContent.trim() === '403 Forbidden') {
+        showErrorView(
+          sk,
+          403,
+          i18n(sk, 'error_site_403'),
+          i18n(sk, 'user_switch'),
+          () => {
+            login(sk, true); // select account
+            sk.addEventListener('loggedin', () => {
+              window.location.reload();
+            }, { once: true });
+          },
+        );
+      }
+    }
+  }
+
+  /**
    * The sidekick provides helper tools for authors.
    * @augments HTMLElement
    */
@@ -3266,6 +3365,7 @@ import sampleRUM from './rum.js';
         checkPlugins(this);
         checkLastModified(this);
         enableInfoBtn(this);
+        handleSiteError(this);
       });
       this.addEventListener('shown', async () => {
         pushDownContent(this);